requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic
python-dotenv==1.1.1
zai-sdk==0.0.3.3